requirements.txt
Pillow==7.2.0
jsonlib-python3==1.6.1
